requirements.txt
streamlit
mysql-connector-python
google-api-python-client
pandas
numpy
scikit-learn
nltk
Sastrawi
python-dotenv